Create ping_sweep.py
Ping sweeping is used to determine the reachability of multiple hosts within a given IP range. 
it is a simple and quick way to identify active hosts on a network.
How to Use:
Call ping_sweep() with a subnet (without the last octet) as the argument, like so:

ping_sweep("192.168.1")

This will check all IPs from 192.168.1.1 to 192.168.1.254.

Note:
If you run this on a large subnet, it may take a while because you're pinging each host individually.
You might want to consider adding concurrency or multi-threading if performance becomes an issue on larger networks.